Add Influence cache and reorganize time server
- No longer sends influence updates every time a map server updates it upstream
- Send influence updates periodically via time server
- Move time server to a separate file
- Move gobbie box logic to a separate file
