+++ src/world/world_server.cpp
@@ -22,132 +22,17 @@ along with this program.  If not, see http://www.gnu.org/licenses/
 
 #include "common/application.h"
 #include "common/logging.h"
-#include "common/vana_time.h"
-
-namespace
-{
-    std::unique_ptr<SqlConnection>  sql;
-    std::shared_ptr<ConquestSystem> conquest;
-} // namespace
-
-void UpdateDailyTallyPoints()
-{
-    uint16 dailyTallyLimit  = settings::get<uint16>("main.DAILY_TALLY_LIMIT");
-    uint16 dailyTallyAmount = settings::get<uint16>("main.DAILY_TALLY_AMOUNT");
-
-    const char* fmtQuery = "UPDATE char_points \
-            SET char_points.daily_tally = LEAST(%u, char_points.daily_tally + %u) \
-            WHERE char_points.daily_tally > -1;";
-
-    int32 ret = sql->Query(fmtQuery, dailyTallyLimit, dailyTallyAmount);
-
-    if (ret == SQL_ERROR)
-    {
-        ShowError("Failed to update daily tally points");
-    }
-    else
-    {
-        ShowDebug("Distributed daily tally points");
-    }
-
-    fmtQuery = "DELETE FROM char_vars WHERE varname = 'gobbieBoxUsed';";
-
-    if (sql->Query(fmtQuery, dailyTallyAmount) == SQL_ERROR)
-    {
-        ShowError("Failed to delete daily tally char_vars entries");
-    }
-}
-
-int32 time_server(time_point tick, CTaskMgr::CTask* PTask)
-{
-    TracyZoneScoped;
-    TIMETYPE VanadielTOTD = CVanaTime::getInstance()->SyncTime();
-
-    // Weekly update for conquest (sunday at midnight)
-    static time_point lastConquestTally  = tick - 1h;
-    static time_point lastConquestUpdate = tick - 1h;
-    if (CVanaTime::getInstance()->getJstWeekDay() == 1 && CVanaTime::getInstance()->getJstHour() == 0 && CVanaTime::getInstance()->getJstMinute() == 0)
-    {
-        if (tick > (lastConquestTally + 1h))
-        {
-            conquest->updateWeekConquest();
-            lastConquestTally = tick;
-        }
-    }
-    // Hourly conquest update
-    else if (CVanaTime::getInstance()->getJstMinute() == 0)
-    {
-        if (tick > (lastConquestUpdate + 1h))
-        {
-            conquest->updateHourlyConquest();
-            lastConquestUpdate = tick;
-        }
-    }
-
-    // Vanadiel Hour
-    static time_point lastVHourlyUpdate = tick - 4800ms;
-    if (CVanaTime::getInstance()->getMinute() == 0)
-    {
-        if (tick > (lastVHourlyUpdate + 4800ms))
-        {
-            lastVHourlyUpdate = tick;
-        }
-    }
-
-    // JST Midnight
-    static time_point lastTickedJstMidnight = tick - 1h;
-    if (CVanaTime::getInstance()->getJstHour() == 0 && CVanaTime::getInstance()->getJstMinute() == 0)
-    {
-        if (tick > (lastTickedJstMidnight + 1h))
-        {
-            if (settings::get<bool>("main.ENABLE_DAILY_TALLY"))
-            {
-                UpdateDailyTallyPoints();
-            }
-
-            lastTickedJstMidnight = tick;
-        }
-    }
-
-    // 4-hour RoE Timed blocks
-    static time_point lastTickedRoeBlock = tick - 1h;
-    if (CVanaTime::getInstance()->getJstHour() % 4 == 0 && CVanaTime::getInstance()->getJstMinute() == 0)
-    {
-        if (tick > (lastTickedRoeBlock + 1h))
-        {
-            lastTickedRoeBlock = tick;
-        }
-    }
-
-    // Vanadiel Day
-    static time_point lastVDailyUpdate = tick - 4800ms;
-    if (CVanaTime::getInstance()->getHour() == 0 && CVanaTime::getInstance()->getMinute() == 0)
-    {
-        if (tick > (lastVDailyUpdate + 4800ms))
-        {
-            lastVDailyUpdate = tick;
-        }
-    }
-
-    if (VanadielTOTD != TIME_NONE)
-    {
-    }
-
-    return 0;
-}
+#include "time_server.h"
 
 WorldServer::WorldServer(int argc, char** argv)
 : Application("world", argc, argv)
-, conquestSystem(std::make_shared<ConquestSystem>())
+, sql(std::make_unique<SqlConnection>())
+, conquestSystem(std::make_unique<ConquestSystem>())
 , httpServer(std::make_unique<HTTPServer>())
 , messageServer(std::make_unique<message_server_wrapper_t>(std::ref(m_RequestExit)))
 {
-    // Anonymous namespace preparation
-    sql      = std::make_unique<SqlConnection>();
-    conquest = conquestSystem;
-
     // Tasks
-    CTaskMgr::getInstance()->AddTask("time_server", server_clock::now(), nullptr, CTaskMgr::TASK_INTERVAL, time_server, 2400ms);
+    CTaskMgr::getInstance()->AddTask("time_server", server_clock::now(), this, CTaskMgr::TASK_INTERVAL, time_server, 2400ms);
 }
 
 WorldServer::~WorldServer() = default;
